+++ src/main/java/menu/domain/CantEatingMenus.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.domain;
 
 import java.util.ArrayList;
 import java.util.Arrays;
