+++ src/main/java/menu/domain/Category.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.domain;
 
 import java.util.Arrays;
 import java.util.List;
