+++ src/main/java/menu/domain/Coach.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.domain;
 
 public class Coach {
     private final String coachName;
