+++ src/main/java/menu/domain/Coaches.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.domain;
 
 import java.util.ArrayList;
 import java.util.List;
