+++ src/test/java/menu/domain/CantEatingMenusTest.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.domain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.*;
