+++ src/test/java/menu/domain/MenuTest.java
@@ -1,4 +1,4 @@
-package menu;
+package menu.domain;
 
 import static org.assertj.core.api.Assertions.*;
